refactor compliance plugin
